Move user ID cookie retrieval from db to server
